alteração na inserção de registros e nos formulários de cadastro para uso do tipo de requisição POST.

+++ storage/databases/db_ecommerce.sql
@@ -18,7 +18,7 @@ CREATE TABLE categorias(
 CREATE TABLE produtos(
   id_produto INT NOT NULL AUTO_INCREMENT PRIMARY KEY ,
   nome VARCHAR(100) NOT NULL ,
-  preco DOUBLE(10,2) NOT NULL ,
+  preco DECIMAL (10,2) NOT NULL ,
   descricao LONGTEXT NULL ,
   categoria_id INT NOT NULL ,
   FOREIGN KEY(categoria_id) REFERENCES categorias(id_categoria)
